pyproject.toml
[project]
name = "langmem0"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1,<2",
    "langchain-huggingface~=1.0",
    "langchain-openai>=1.1.6",
    "langgraph>=1,<2",
    "langgraph-checkpoint>=3.0.1",
    "mem0ai>=1.0.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
]

[tool.setuptools]
#packages = ["langmem0"]
#package-dir = {"" = "src"}
packages.find.where = ["src"]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
    "sphinx>=9.1.0",
    "sphinx-autodoc-typehints>=3.6.2",
]
